requirements.txt
absl-py
aiohttp
aiosignal
annotated-types
anyio
asgiref
attrs
beautifulsoup4
blinker
certifi
cffi
charset-normalizer
click
contourpy
crispy-bootstrap5
cryptography
cycler
deprecation
dj-database-url
Django
django-bootstrap5
django-cors-headers
django-crispy-forms
django-email
django-environ
django-extensions
django-filter
django-humanize
djangorestframework
djangorestframework_simplejwt
drf-yasg
filelock
flatbuffers
fonttools
frozenlist
gotrue
gunicorn
h11
h2
hpack
httpcore
httpx
humanize
hyperframe
idna
inflection
itsdangerous
Jinja2
joblib
kiwisolver
Markdown
MarkupSafe
matplotlib
mediapipe
ml_dtypes
multidict
numpy
opencv-contrib-python
opencv-python
opt_einsum
packaging
pillow
postgrest
protobuf
psycopg2-binary
pydantic
pydantic_core
pygame
PyJWT
Pympler
pyparsing
PyPDF2
python-dateutil
pyttsx3
pytz
pywhatkit
qrcode
realtime
requests
scikit-learn
scipy
sentencepiece
setuptools
six
sniffio
sounddevice
soupsieve
sqlparse
sslcommerz-lib
storage3
StrEnum
supabase
supafunc
threadpoolctl
typing_extensions
tzdata
uritemplate
urllib3
websockets
Werkzeug
whitenoise
wikipedia
yarl
